# Performance of Interpreter construction