The `Interpreter` is constructed many times when fitting with `lmfit`. This PR improves performance by

* Taking the calculation of the number version out of the loop
* Changing the order of the check for `no_deepcopy`

Current master branch:
```
%timeit Interpreter()
364 µs ± 21.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
``` 
With this PR
``` 
%timeit Interpreter()
271 µs ± 23.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
``` 

@newville