@newville 2022-05-23
@eendebakpt thanks! I'll merge this